converted BigInteger to long

+++ meteor/Converters/LineCountToHeightConverter.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Numerics;
 using Avalonia.Data.Converters;
 
 namespace meteor.Converters;
@@ -11,11 +10,11 @@ public class LineCountToHeightConverter : IMultiValueConverter
     public object? Convert(IList<object?> values, Type targetType, object? parameter, CultureInfo culture)
     {
         const int verticalPadding = 5;
-        if (values is [BigInteger lineCount, double lineHeight, double minHeight])
+        if (values is [long lineCount, double lineHeight, double minHeight])
         {
             // Calculate the height using the adjusted line height
             var adjustedLineHeight = lineHeight * 1.5;
-            var calculatedHeight = (double)(lineCount * (BigInteger)adjustedLineHeight);
+            var calculatedHeight = lineCount * adjustedLineHeight;
             // Hacky method to fix text rendering below the viewport
             calculatedHeight = (double)lineCount * lineHeight;
             return Math.Max(calculatedHeight + verticalPadding, minHeight);
