+++ meteor/Services/CursorPositionService.cs
@@ -1,14 +1,13 @@
 using System;
-using System.Numerics;
 using meteor.Interfaces;
 
 namespace meteor.Services;
 
 public class CursorPositionService : ICursorPositionService
 {
-    public event Action<BigInteger, BigInteger[]>? CursorPositionChanged;
+    public event Action<long, long[]>? CursorPositionChanged;
 
-    public void UpdateCursorPosition(BigInteger position, BigInteger[] lineStarts)
+    public void UpdateCursorPosition(long position, long[] lineStarts)
     {
         CursorPositionChanged?.Invoke(position, lineStarts);
     }
